pyproject.toml
[tool.poetry]
name = "paypayopasdk"
version = "0.1.0"
description = ""
authors = ["TakayukiKomada <Takakoma.79@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
paypayopa = "^1.7"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
